+++ python/petrify/__init__.py
@@ -49,4 +49,4 @@ __all__ = [
     "verify_equal_area",
 ]
 
-__version__ = "0.12.0"
+__version__ = "0.13.0"